requirements.txt
streamlit
markdown2
pdfplumber
PyPDF2
openai
python-dotenv
PyMuPDF
